+++ backend/src/Logitar.Faktur.Domain/Banners/Events/BannerCreatedEvent.cs
@@ -5,11 +5,11 @@ namespace Logitar.Faktur.Domain.Banners.Events;
 
 public record BannerCreatedEvent : DomainEvent, INotification
 {
+  public DisplayName DisplayName { get; init; }
+
   public BannerCreatedEvent(ActorId actorId, DisplayName displayName)
   {
     ActorId = actorId;
     DisplayName = displayName;
   }
-
-  public DisplayName DisplayName { get; init; }
 }