+++ backend/src/Logitar.Faktur.Domain/Stores/Events/StoreCreatedEvent.cs
@@ -5,11 +5,11 @@ namespace Logitar.Faktur.Domain.Stores.Events;
 
 public record StoreCreatedEvent : DomainEvent, INotification
 {
+  public DisplayName DisplayName { get; init; }
+
   public StoreCreatedEvent(ActorId actorId, DisplayName displayName)
   {
     ActorId = actorId;
     DisplayName = displayName;
   }
-
-  public DisplayName DisplayName { get; init; }
 }