+++ backend/src/Logitar.Faktur.EntityFrameworkCore.Relational/ISqlHelper.cs
@@ -5,6 +5,7 @@ namespace Logitar.Faktur.EntityFrameworkCore.Relational;
 
 public interface ISqlHelper
 {
+  IQueryBuilder ApplyIdSearch(IQueryBuilder builder, TextSearch search, ColumnId column);
   IQueryBuilder ApplyTextSearch(IQueryBuilder builder, TextSearch search, params ColumnId[] columns);
   IQueryBuilder QueryFrom(TableId table);
 }