+++ backend/src/Logitar.Faktur.Web/Models/Stores/SearchBannersQuery.cs
@@ -13,11 +13,12 @@ public record SearchStoresQuery : SearchQuery
   {
     SearchStoresPayload payload = new()
     {
-      IdIn = IdIn,
       BannerId = BannerId,
       Skip = Skip,
       Limit = Limit
     };
+    payload.Id.Operator = IdOperator;
+    payload.Id.Terms = IdTerms.Select(value => new SearchTerm(value));
     payload.Search.Operator = SearchOperator;
     payload.Search.Terms = SearchTerms.Select(value => new SearchTerm(value));